Fixes a potential race condition from two callers
It was possible for two callers to hit an interesting race condition
where:

If two callers called into the runner at the same time and end up
needing to hit the upstream, both would enter the CoroutineRunner at the
same point.

At this point one would claim the mutex while the other waits. The
caller A would check the active task, see that there is none, and
continue.

At this point caller A would create a new task.
Caller B would enter the mutex at the same point at which caller A is
creating a new task.

Caller A would then attempt to claim the lock again to set itself as the
active task. Caller B would be holding the lock as it attempts to check
for an active task, of which there is none. Once caller B leaves the
lock, caller A would claim it and set itself active, but caller B would
at this point be creating a new task and waiting to claim the lock to
set itself active - thus creating a situation where two calls to the
upstream are possible at once.

By wrapping the entire block of check -> create -> mark active in the
same single mutex lock block, this race should be avoided as whoever
claims the mutex first will become the active for whoever comes next.